refactor(core): 收敛 core→product 模块级依赖倒置
将 12 处模块级反向 import 改为惰性/局部/TYPE_CHECKING 依赖，仅保留 boot 与 ai_callback/runner 两处合理接线；新增 federate/deployment.py 惰性部署标识代理。110 个相关测试通过。
